fix(#3723): port #3612's raw_value.rs changes across the epic #1116 split
The rebase onto #3612 (PR #3736) conflicted on
`row_decoder/raw_value.rs`: this branch SPLIT that file into a
`raw_value/` directory module while #3612 MODIFIED it in place. Keeping
either side wholesale loses something, so the deletion is kept and all
three of #3612's changes are ported into the split:

1. The `float` arm returns `Value::Float32(f)`, not the f64
   `Value::Float(f as f64)` (raw_value/fixed_width.rs). This is the port
   that would have silently regressed: the f32->f64 widening is lossless
   for the values a test picks, so the old assertion passed either way.

2. `primitive_marshal_to_cql_short` keeps #3612's EFFECTIVE visibility,
   not its spelling. #3612 widened it to `pub(super)` from
   `raw_value.rs`, a child of `row_decoder`; after the split the item
   sits in `raw_value/marshal_short.rs`, a child of `raw_value`, where
   the same token means something NARROWER. Widened explicitly to
   `pub(in ...::row_decoder)` so #3612's caller in
   `complex_column::cell_path_key` still reaches it (verified: the
   `pub(super)` spelling fails to compile, E0624).

3. The test pinning the widening asserts `Value::Float32(1.5)`.

Adds a pin asserting the float DISCRIMINANT (not the magnitude) on all
three routes into the fixed-width arms after the split, including the
nested-collection path #3723's width guards sit on.

+++ cqlite-core/src/storage/sstable/reader/parsing/row_decoder/raw_value/fixed_width.rs
@@ -162,8 +162,10 @@ impl V5CompressedLegacyParser {
                 Ok(Value::Uuid(uuid))
             }
             "float" => {
+                // CQL `float` is `Value::Float32`, not the f64 `Value::Float`; the column
+                // path and both UDT field decoders already agree (roborev round 10 F1).
                 let f = f32::from_be_bytes([data[0], data[1], data[2], data[3]]);
-                Ok(Value::Float(f as f64))
+                Ok(Value::Float32(f))
             }
             "double" => Ok(Value::Float(f64::from_be_bytes([
                 data[0], data[1], data[2], data[3], data[4], data[5], data[6], data[7],
